refactor: move resources inside engine, make material struct
