+++ common/src/collision/collidable.rs
@@ -19,7 +19,7 @@ pub struct Sweep<T> {
     pub velocity: Vector3<f32>,
 }
 
-impl <T> Sweep<T> {
+impl<T> Sweep<T> {
     pub fn new(object: T, velocity: Vector3<f32>) -> Self {
         Self { object, velocity }
     }
