+++ common/src/engine/renderer.rs
@@ -5,14 +5,14 @@ use crate::camera::{Camera, OrbitalCamera};
 use crate::colors::{self, Color, ColorExt};
 use crate::debug::Cuboid;
 use crate::engine::input::Input;
+use crate::engine::resources::{CubemapHandle, TextureHandle};
+use crate::engine::resources::{GeometryHandle, Resources};
 use crate::geometry::primitives;
 use crate::geometry::primitives::SimplePoint;
 use crate::light::Light;
 use crate::line::{Line, LinePoint};
 use crate::maths::Matrix4Ext;
 use crate::quad::QuadVertex;
-use crate::engine::resources::{CubemapHandle, TextureHandle};
-use crate::engine::resources::{GeometryHandle, Resources};
 use crate::world::{QuadBatches, Renderables, World, WorldGraph};
 use crate::{context, maths};
 use color_eyre::Result;
