+++ common/src/lib.rs
@@ -11,10 +11,9 @@ pub mod import;
 pub mod light;
 pub mod line;
 pub mod maths;
-pub mod physics;
 pub mod quad;
 pub mod run;
 pub mod serde;
 pub mod terrain;
-pub mod texture;
+pub mod material;
 pub mod world;
