+++ common/src/material/mod.rs
@@ -4,4 +4,5 @@ pub use texture2d::Texture2DResource;
 mod texture;
 
 pub mod cubemap;
+mod material;
 pub mod texture2d;
