+++ common/src/material/texture.rs
