+++ common/src/world/world.rs
@@ -1,10 +1,10 @@
 use crate::collision::collidable::{RayHitNode, Sweep, SweepHitNode};
 use crate::collision::colliders::sphere::Sphere;
 use crate::engine::renderer::{Background, Renderable};
+use crate::engine::resources::Resources;
 use crate::light::Light;
 use crate::line::Line;
 use crate::maths::Ray;
-use crate::engine::resources::Resources;
 use crate::serde::SerializedWorld;
 use crate::world::graph::WorldGraph;
 use crate::world::{PhysicsContext, QuadTree};
