+++ tests/src/common/colliders/aabb_tests.rs
@@ -4,8 +4,8 @@ mod tests {
     use common::collision::collidable::{BroadPhaseCollisionQuery, NarrowPhaseCollisionQuery};
     use common::collision::colliders::aabb::Aabb;
     use common::collision::colliders::capsule::Capsule;
-    use common::maths::{Local, Ray};
     use common::engine::resources::Resources;
+    use common::maths::{Local, Ray};
     use nalgebra::{Point3, Vector3};
 
     #[test]
