+++ tests/src/util.rs
@@ -1,6 +1,6 @@
-use common::geometry::Geometry;
-use std::path::{Path, PathBuf};
 use common::engine::resources::{GeometryHandle, Resources};
+use common::geometry::Geometry;
+use std::path::Path;
 
 pub fn load_test_geometry_handle(path: &Path, resources: &mut Resources) -> GeometryHandle {
     resources.get_geometry_handles(path, None).unwrap()[0]
